Final refactor.
